Separation of unit and integration test.  The latter is dependent on existing of a Docker container
